Refactor: Remove AdminLayout component from admin pages
This commit removes the `AdminLayout` component from several admin pages, including:

- Admin
- EditCategory
- EditProduct
- NewProduct
- NewCategory
- Products

The `AdminLayout` component was previously wrapping the content of these pages. This change simplifies the component structure and potentially allows for more flexible layout options in the future. The `AdminLayout` is now applied within the `App.tsx` file using a `ProtectedRoute` component. This ensures that all admin routes are protected and wrapped with the `AdminLayout`.