Refactor: Remove AdminLayout component from more admin pages
This commit removes the `AdminLayout` component from the following admin pages:

- Categories
- Products
- NewProduct
- EditProduct
- NewCategory
- EditCategory
- Admin

The `AdminLayout` component was previously wrapping the content of these pages. This change simplifies the component structure, as the `AdminLayout` is now applied within the `App.tsx` file using a `ProtectedRoute` component. This ensures that all admin routes are protected and wrapped with the `AdminLayout`.